+++ history.md
@@ -1,5 +1,7 @@
 # History
 
+* 2023.09.21 : LaTeX and Text output of proof, with verbose option  
+  which shows the proof validation result.
 * 2023.09.19 : Proof verifier for Propositional logic
 * 2023.09.12 : Parser for Fitch proofs
 * 2023.08.09 : Document.pdf